Chore: Create hook to handle api users data
